Cover the timeline cursor's remaining endings
The cursor's grab had four endings only a unit test can reach, and one
branch nothing could reach at all.

`scrub` carried its own copy of the match that turns a finished run into
an outcome, so the two could come to disagree about what a failed run
is. Split `grabbed` into the caching half and `drawn_frame`, which both
paths now share — caching is then the only difference between them.

`move_cursor` took an `i64` and refused a step wider than a `u32`, which
no caller can ask for: the keys pass one and ten. Taking an `i32`
deletes the branch rather than testing it.

The rest is tests: staging that cannot be written, a page that closes
mid-grab, a dropped channel, a timeline with its cues taken away, and
the cursor frame's `Debug`.

Verified with cargo fmt, both clippy invocations, 1292 unit tests and
the full 52-scenario e2e suite. Merged coverage 97.36% line / 83.95%
branch; subtitle_edit.rs 99.57%/95.08%, preview.rs 98.19%/79.78%.

+++ src/app.rs
@@ -3166,7 +3166,7 @@ impl App {
 
     /// Moves the timeline cursor by `steps` of `subtitle_edit::TIMELINE_STEP`, for `h`/`l`
     /// and `H`/`L` while the timeline holds the cursor.
-    pub fn move_timeline_cursor(&mut self, steps: i64) {
+    pub fn move_timeline_cursor(&mut self, steps: i32) {
         if let Some(state) = self.subtitle_edit.as_mut()
             && state.move_cursor(steps)
         {
